Add network locality fields
This adds `network.locality` that simply has a value of either private or public. If either `source.ip` or `destination.ip` are public IP addresses then network.locality is elevated to "public". Otherwise if both `source.ip` and `destination.ip` are non-public then `network.locality` is private.

The IPv4 and IPv6 ranges that are considered private are strictly specified in the definition of `network.locality`.

This is a useful means of filtering on flows. Some common queries I use are

- network.locality:public and source.locality:public
- network.locality:public and destination.locality:public